Flyttet gjenbrukbare funksjoner til common.js
